+++ src/control/mod.rs
@@ -1,0 +1,6 @@
+pub use self::character_controller::{
+    CharacterAutostep, CharacterCollision, CharacterLength, KinematicCharacterController,
+    KinematicCharacterControllerOutput, MoveShapeOptions, MoveShapeOutput,
+};
+
+mod character_controller;