+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.17.0 (02 Oct. 2022)
+### Added
+- Add a **kinematic character controller** implementation. This feature is accessible in two different ways:
+  1. The first approach is to insert the `KinematicCharacterController` component to an entity. If the
+  `KinematicCharacterController::custom_shape` field is set, then this shape is used for the character control.
+  If this field is `None` then the `Collider` attached to the same entity as the character controller is used.
+  The character controller will be automatically updated when the `KinematicCharacterController::movement` is set.
+  The result position is written to the `Transform` of the character controller’s entity.
+  2. The second, lower level, approach, is to call `RapierContext::move_shape` to compute the possible movement
+  of a shape, taking obstacle and sliding into account.
+- Add implementations of `Add`, `AddAssign`, `Sub`, `SubAssign` to `ExternalForce` and `ExternalImpulse`.
+- Add `ExternalForce::at_point` and `ExternalImpulse::at_point` to apply a force/impulse at a specific point
+  of a rigid-body.
+
+### Fix
+- Fix shapes quickly switching between scaled and non-scaled versions due to rounding errors in the scaling extraction
+  from bevy’s global affine transform.
+
 ## 0.16.2 (23 August 2022)
 ### Added
 - Implement `Debug` for `Collider` and `ColliderView`.